+++ docs/whats_new.rst
@@ -23,6 +23,7 @@ New Features
 Internal Changes
 ~~~~~~~~~~~~~~~~
 
+- Update vc option in preprocessor to read in csv file (:pull:`294`).
 - Update Python version in CI workflow (:pull:`295`).
 - Update docstring (:pull:`280`).
 - Update docstrings (:pull:`277`).